+++ CHANGELOG.md
@@ -18,6 +18,7 @@ NOTE: Breaking release in controllers.
 - Refactor metrics recorder implementation including the prometheus backend.
 - Refactor internal controller queue into a decorator implementation approach.
 - Remove `Delete` method from `controller.Handler` and simplify to only `Handle` method
+- Add `DisableResync` flag on controller configuration to disable the resync of all resources.
 
 ## [0.8.0] - 2019-12-11